set boosted yields percentage fix
