+++ energy-integration/farm-boosted-yields/src/lib.rs
@@ -44,14 +44,6 @@ pub trait FarmBoostedYieldsModule:
     + weekly_rewards_splitting::update_claim_progress_energy::UpdateClaimProgressEnergyModule
     + energy_query::EnergyQueryModule
 {
-    #[endpoint(setBoostedYieldsRewardsPercentage)]
-    fn set_boosted_yields_rewards_percentage(&self, percentage: u64) {
-        self.require_caller_has_admin_permissions();
-        require!(percentage <= MAX_PERCENT, "Invalid percentage");
-
-        self.boosted_yields_rewards_percentage().set(percentage);
-    }
-
     #[endpoint(collectUndistributedBoostedRewards)]
     fn collect_undistributed_boosted_rewards(&self) {
         self.require_caller_has_admin_permissions();
