+++ farm-staking/farm-staking/wasm/src/lib.rs
@@ -22,6 +22,7 @@ multiversx_sc_wasm_adapter::endpoints! {
         init => init
         upgrade => upgrade
         mergeFarmTokens => merge_farm_tokens_endpoint
+        setBoostedYieldsRewardsPercentage => set_boosted_yields_rewards_percentage
         calculateRewardsForGivenPosition => calculate_rewards_for_given_position
         topUpRewards => top_up_rewards
         withdrawRewards => withdraw_rewards
@@ -70,7 +71,6 @@ multiversx_sc_wasm_adapter::endpoints! {
         unstakeFarmThroughProxy => unstake_farm_through_proxy
         unbondFarm => unbond_farm
         claimBoostedRewards => claim_boosted_rewards
-        setBoostedYieldsRewardsPercentage => set_boosted_yields_rewards_percentage
         collectUndistributedBoostedRewards => collect_undistributed_boosted_rewards
         getBoostedYieldsRewardsPercentage => boosted_yields_rewards_percentage
         getAccumulatedRewardsForWeek => accumulated_rewards_for_week
